refactor(homepage): move components to folder
