+++ app/components/homepage/LatestNewsCards.tsx
@@ -1,4 +1,3 @@
-'use client'
 import React, { useState, useEffect } from 'react'
 import { useSession } from 'next-auth/react'
 
@@ -14,7 +13,8 @@ import { useChosenTagsContext } from '@/context/ChosenTagsContext'
 
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar'
 import { Card, CardHeader, CardTitle } from '@/components/ui/card'
-import { Skeleton } from '@/components/ui/skeleton'
+
+import LatestNewsCardsSkeleton from './LatestNewsCardsSkeleton'
 
 // const stories = [
 //   {
@@ -115,35 +115,7 @@ export default function LatestNewsCards({ className }: { className: string }) {
   return (
     <section className={className}>
       <ul className="space-y-12 md:space-y-8">
-        {isLoading && (
-          <Card className="h-96 lg:h-48 grid grid-cols-1 lg:grid-cols-2 grid-rows-5 lg:grid-rows-1 justify-between p-2 gap-4 bg-white bg-opacity-50 rounded-md">
-            <CardHeader className="row-span-2 h-fit flex flex-col gap-x-2 gap-y-4 p-2 justify-between">
-              <article className="grid gap-y-2">
-                <section className="flex items-center gap-x-2">
-                  <Skeleton className="size-8 rounded-full" />
-
-                  <Skeleton className="w-24 h-fit">
-                    <div className="invisible">invisible</div>
-                  </Skeleton>
-
-                  <Skeleton className="w-24 h-8 block lg:hidden ml-auto" />
-                </section>
-
-                <Skeleton className="w-full h-fit text-transparent text-xl font-bold">
-                  <div className="invisible">invisible</div>
-                  <div className="invisible">invisible</div>
-                  <div className="invisible">invisible</div>
-                </Skeleton>
-              </article>
-
-              <Skeleton className="w-24 h-fit hidden lg:block text-xs">
-                <div className="invisible">invisible</div>
-              </Skeleton>
-            </CardHeader>
-
-            <Skeleton className="row-span-3 size-full ml-auto rounded-2xl" />
-          </Card>
-        )}
+        {isLoading && <LatestNewsCardsSkeleton />}
 
         {selectedStories?.map((story) => {
           const { id, title, author, tags, createdAt } = story
