+++ app/components/homepage/index.tsx
@@ -1,0 +1,4 @@
+import LatestNews from './LatestNews'
+import MainCarousel from './MainCarousel'
+
+export { LatestNews, MainCarousel }
